# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Arc Pass API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /passes/me:
    get:
      operationId: listMyPasses
      tags: [passes]
      summary: Get the current user's Founder and Builder passes
      responses:
        "200":
          description: My passes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyPasses"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /passes/founder/{id}:
    get:
      operationId: getFounderPass
      tags: [passes]
      summary: Get a public minted Founder Pass by ID
      description: Returns 404 until the credential is minted and intentionally public.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Founder Pass
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FounderPass"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /passes/founder/{id}/download-url:
    get:
      operationId: getFounderPassDownloadUrl
      tags: [passes]
      summary: Get a signed download URL for a claimed Founder Pass image
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Download URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DownloadUrlResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /passes/founder/claim:
    post:
      operationId: claimFounderPass
      tags: [passes]
      summary: Claim the current user's eligible Founder Pass
      responses:
        "200":
          description: Claimed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FounderPass"
        "400":
          description: Not eligible or already claimed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /passes/founder/mint:
    post:
      operationId: mintFounderPass
      tags: [passes]
      summary: Mint the current user's claimed Founder Pass onchain
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MintRequest"
      responses:
        "200":
          description: Minted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FounderPass"
        "400":
          description: Not claimable or already minted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /passes/builder/{id}:
    get:
      operationId: getBuilderPass
      tags: [passes]
      summary: Get a public minted Onchain Builder Pass by ID
      description: Returns 404 until the credential is minted and intentionally public.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Builder Pass
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BuilderPass"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /passes/builder/{id}/download-url:
    get:
      operationId: getBuilderPassDownloadUrl
      tags: [passes]
      summary: Get a signed download URL for a claimed Builder Pass image
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Download URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DownloadUrlResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /passes/builder/verify:
    post:
      operationId: verifyBuilder
      tags: [passes]
      summary: Run initial Builder verification using connected accounts and wallets
      description: >
        Requires Discord connected (required), GitHub connected (required),
        and at least one wallet added. Analyzes deterministic onchain data
        plus qualitative GitHub signal and assigns the highest qualifying
        tier. Requires at least one valid deployed contract to qualify for
        any tier.
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BuilderVerificationResult"
        "400":
          description: Missing required connections
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /passes/builder/reverify:
    post:
      operationId: reverifyBuilder
      tags: [passes]
      summary: Re-run Builder verification to check for a tier upgrade
      description: >
        Enforces a 7-day cooldown since the last verification. Computes a
        proposed tier but does not apply it — call POST
        /passes/builder/upgrade to confirm and apply.
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BuilderVerificationResult"
        "400":
          description: No claimed Builder Pass to re-verify
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Re-verification cooldown active
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CooldownErrorResponse"
  /passes/builder/upgrade:
    post:
      operationId: upgradeBuilderTier
      tags: [passes]
      summary: Confirm and apply the most recently proposed tier upgrade
      responses:
        "200":
          description: Upgraded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BuilderPass"
        "400":
          description: No pending upgrade available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /passes/builder/claim:
    post:
      operationId: claimBuilderPass
      tags: [passes]
      summary: Claim the current user's verified Builder Pass
      responses:
        "200":
          description: Claimed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BuilderPass"
        "400":
          description: Not eligible or already claimed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /passes/builder/mint:
    post:
      operationId: mintBuilderPass
      tags: [passes]
      summary: Mint the current user's claimed Builder Pass onchain
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MintRequest"
      responses:
        "200":
          description: Minted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BuilderPass"
        "400":
          description: Not claimable, already minted, or supply exhausted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [passes]
      summary: Get dashboard summary stats for the current user
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    CooldownErrorResponse:
      type: object
      required: [error, nextVerificationAt]
      properties:
        error:
          type: string
        nextVerificationAt:
          type: string
          format: date-time
...
    FounderPass:
      type: object
      required: [id, variant, eligibilityStatus, claimStatus, createdAt]
      properties:
        id:
          type: integer
        variant:
          type: string
          enum: [normal, premium_black]
        founderTier:
          oneOf:
            - $ref: "#/components/schemas/FounderTier"
            - type: "null"
        displayName:
          type: ["string", "null"]
        username:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        founderTitle:
          type: ["string", "null"]
        companyName:
          type: ["string", "null"]
        companyIndustry:
          type: ["string", "null"]
        companyLogoUrl:
          type: ["string", "null"]
        companyWebsite:
          type: ["string", "null"]
        companyLocation:
          type: ["string", "null"]
        startupStage:
          type: ["string", "null"]
        founderStatement:
          type: ["string", "null"]
        companyDescription:
          type: ["string", "null"]
        eligibilityStatus:
          type: string
          enum: [eligible, invite_required, under_review, ineligible]
        claimStatus:
          type: string
          enum: [locked, claimed, minted]
        passNumber:
          type: ["integer", "null"]
        network:
          type: ["string", "null"]
          enum: [arc, base, null]
        tokenId:
          type: ["string", "null"]
        contractAddress:
          type: ["string", "null"]
        destinationWallet:
          type: ["string", "null"]
        transactionHash:
          type: ["string", "null"]
        issuedAt:
          type: ["string", "null"]
          format: date-time
        claimedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    MyPasses:
      type: object
      properties:
        founder:
          oneOf:
            - $ref: "#/components/schemas/FounderPass"
            - type: "null"
        builder:
          oneOf:
            - $ref: "#/components/schemas/BuilderPass"
            - type: "null"
...
    MintRequest:
      type: object
      required: [mintMethod]
      properties:
        mintMethod:
          type: string
          enum: [wallet_connect, manual_address]
        walletAddress:
          type: string
        network:
          type: string
          enum: [arc, base]
          default: "arc"
        confirmed:
          type: boolean
          description: Required true for manual_address — confirms the user reviewed the warning
    DownloadUrlResponse:
      type: object
      required: [url]
      properties:
        url:
          type: string
...